Use Github as download source instead of imagemagick.org
It looks like each time a new release is published that the links to the previous version downloads become broken, resulting in a build error for pipeline apps that don't already have a cached ImageMagick installed. I think the Github source here is permanently-accessible so this should allow me to continue using x.25 (or whatever I land on) even as additional releases are published.